Add a Minmax worker
The minmax worker is a daemon, separate from the main program, whose job is to recompute daily climatology in the past. When a connector in the main program recovers data dating from some time, it schedules a "job", a request that climatology is recomputed for the time range newly recovered. That job is picked up by the minmax worker and climatology is recomputed.