LSN50v2Probe6470: Use a more precise resistance
The LSN50v2 uses an ADC to measure a temperature through a thermistance.
This ADC actually measures the voltage across a voltage divider so the
value measured depends on the value of a resistance soldered across two
pins. Since the resistances are not usually very accurate (+/- 10% of
the nominal value), the device's firmware also sends a more exact value
of the resistance measured by a separate mean and entered into the
device configuration. We now use that value instead of the nominal
10kOhms.